revenue-distribution: split distribution configuration (#17)
- split configure-distribution instruction instruction into
  - configure-distribution-payments
  - configure-distribution-rewards
  - finalize-distribution-rewards

The existing instruction was awkward because it checks two types of
authorities: payments accountant or rewards accountant. This splits the
logic to make things cleaner.

This change also adds some specific criteria to finalize the
distribution rewards and makes this action permissionless. The critera
are:
- must be X epochs from the distribution's defined epoch (where X is
fetched from the program config)
- payments must be finalized
- rewards have not been finalized yet

The distribution has a new field tracking uncollectible SOL payments in
case the payments accountant needs to write off SOL.

Closes https://github.com/malbeclabs/doublezero/issues/1033.
Closes https://github.com/malbeclabs/doublezero/issues/1046.