reward-distribution: add reward distribution precursor instructions (#45)
This PR adds instructions required to distribute rewards to network
contributors. Technically the reward share verification is not
necessary, but it provides a way for the rewards accountant to validate
the merkle root it posted for rewards.

**The test to sweep distribution tokens is ignored when not in the
development feature.** This is because there is no swap program that the
Revenue Distribution program can integrate with yet. In lieu of using a
swap program, a fixed swap rate will be used. We can then test
end-to-end with Solana program tests and on Solana devnet by
transferring 2Z to the swap destination token account for sweeps.

- Add reward share verification
- This change specifies the reward share leaf for the rewards Merkle
tree.
- Each leaf supports reward blocking, which for distributing rewards
will assume the flag is not set (so if it is set for a given contributor
when computing the merkle root, rewards would not be distributed to this
contributor).
- Add initialize-swap-destination instruction.
- This token account will be where the swap program transfers 2Z after
it performs a swap with SOL.
  - Only supported mint is 2Z.
- Add sweep-distribution-tokens instruction.
  - **ONLY FOR feature = "development"**.
- The development instruction processor uses a fixed SOL/2Z swap rate of
100:1.

Distributing rewards will be added in a subsequent PR. **Distributing
rewards will only be supported with feature = "development" until the
swap program is built.**

Closes https://github.com/malbeclabs/doublezero/issues/1341.
Closes https://github.com/malbeclabs/doublezero/issues/1347.