revenue-distribution: distribute rewards (#50)
Distribute rewards to contributor-specified recipients. A portion of
these rewards are burned and the remaining are transferred to recipient
Associated Token Accounts (ATAs).

These recipients are expected to have ATAs already created prior to
invoking this instruction. It will not be the job of the instruction to
create these accounts (although the invoker may create these accounts
prior to distributing rewards).

This PR also removes any mention of claiming because it is not required
that the contributors themselves "claim" rewards since this
distribute-rewards instruction is permissionless (and its invoker will
be paid enough SOL to cover the cost of the transaction to call this
instruction).

**NOTE: This instruction can only be invoked under the development
feature because a prerequisite to calling this instruction is 2Z being
swept into the distribution.** In a subsequent PR that adds a swap
program, we can remove this requirement.

Closes https://github.com/malbeclabs/doublezero/issues/774.
Closes https://github.com/malbeclabs/doublezero/issues/1346.