revenue-distribution: enforce grace period after initializing distribution (#53)
Instead of the calculation grace period being a superficial indicator to
offchain accountants to perform calculations, this change enforces the
grace period by recording the timestamp when calculations are allowed to
be submitted.

This change also restricts the grace period from being set too high
(arbitrarily chose 24 hours as the maximum).

Testing needed to change to use the program test context in order to
manipulate the Clock sysvar account.

Closes: https://github.com/malbeclabs/doublezero/issues/1561.